requirements.inference.txt
torch
torchvision
numpy
pillow
scikit-learn
timm
gradio
transformers
